fix(oauth): check for OAuth error responses even when status code is 200 (#646)
* fix(oauth): check for oauth error responses even when status code is 200

Certain providers (**cough** GitHub **cough**) return HTTP 200 responses
with an error in the body for their OAuth token exchange and
token refresh endpoints. Previously, we only checked for non-200 status
codes, which could lead to us missing the errors, attempting to
unmarshal the body as a token, failing to detect the error, and even
saving the empty token.

This commit attempts to unmarshal the body as an OAuth error response
even when the status code is 200. If the body is not an OAuth error
response, we fall back to the previous behavior.

* Address CodeRabbit's concerns

* Improve test assertions as suggested by CodeRabbit

* Address more CodeRabbit feedback